// src/Ui.Angular/src/app/shared/generated/model/skill-aggregation-level.ts
/**
 * Kaizen.Skills.Service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: latest
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type SkillAggregationLevel = 'None' | 'SkillCategory' | 'Skill' | 'SkillLevel';

export const SkillAggregationLevel = {
    None: 'None' as SkillAggregationLevel,
    SkillCategory: 'SkillCategory' as SkillAggregationLevel,
    Skill: 'Skill' as SkillAggregationLevel,
    SkillLevel: 'SkillLevel' as SkillAggregationLevel
};